LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/range-expected.txt

Input Range


PASS range type support on input element 
PASS min attribute support on input element 
PASS max attribute support on input element 
FAIL Illegal value of min attribute assert_equals: expected "0" but got "ab"
FAIL Illegal value of max attribute assert_equals: expected "100" but got "f"
PASS Converting an illegal string to the default value 
FAIL Converting an illegal string to the default step assert_equals: expected "1" but got "xyz"
PASS the value is set to min when a smaller value than min attribute is given 
PASS the value is set to max when a larger value than max attribute is given 
FAIL default value of min attribute in input type=range assert_equals: expected "0" but got ""
FAIL default value of max attribute in input type=range assert_equals: expected "100" but got ""
PASS default value when min and max attributes are given (= min plus half the difference between min and max) 
PASS default value with step control when both min and max attributes are given 
FAIL default value when both min and max attributes are given, while min > max assert_equals: expected "2" but got "51"
PASS The default step scale factor is 1, unless min attribute has non-integer value 
PASS Step scale factor behavior when min attribute has integer value but max attribute is non-integer  
FAIL The default scale factor is 1 even if step attribute is explicitly set to non-integer value, unless min attribute has non-integer value assert_equals: expected "1" but got "0.5"
PASS Solving the step mismatch 
PASS Performing stepUp() 
PASS Performing stepDown() 
FAIL Performing stepUp() beyond the value of the max attribute The object is in an invalid state.
FAIL Performing stepDown() beyond the value of the min attribute The object is in an invalid state.
